Move methods to domUpdate file
